Improve integration with react-router
This commit improves the overall integration with the
react-router library. A layout component now provides
standard page layout. Basic error handling was added
so that you land on a simple error page with the error
message when something goes wrong. Data loading is now
handled by react-router when a page component loads.