Improve gene search
This commit tweaks the gene search in a couple of minor ways.

First, it adds the model organism database gene ID to the
view used for supporting the gene search. This allows for
database speicifc IDs to be used (e.g. FBgn0000490).

Second, it removes the leading wildcard from the query term
and it sorts by symbol length. This allows for short
but exact matches to rise to the top.

Also adds an `api` target to run scripts.